+++ nym-ip-packet-client/src/lib.rs
@@ -11,4 +11,4 @@ pub use error::Error;
 pub use listener::{IprListener, MixnetMessageOutcome};
 
 // Re-export the currently used version
-pub use nym_ip_packet_requests::v8 as current;
+pub use nym_ip_packet_requests::v9 as current;